test: 프로세스 락 테스트 추가
